moved non-purely client side methods and constants from renderer class to tile entity to prevent potential crashes/unnecessary class loading on servers
